+++ lib/supabase/indexer/indexerState.ts
@@ -1,0 +1,9 @@
+import { RealtimeChannel } from "@supabase/supabase-js";
+
+export type BroadcastHandler = (payload: any) => void;
+
+export const listeners = new Map<string, Set<BroadcastHandler>>();
+
+export const boundEvents = new Set<string>();
+
+export const channelRef = { current: null as RealtimeChannel | null };
